+++ src/main/java/frc/robot/RobotContainer.java
@@ -19,12 +19,12 @@ import edu.wpi.first.wpilibj2.command.RunCommand;
 import edu.wpi.first.wpilibj2.command.button.JoystickButton;
 import edu.wpi.first.wpilibj2.command.button.Trigger;
 import frc.robot.Constants.IntakeConstants;
-import frc.robot.commands.ArmIdleCommand;
-import frc.robot.commands.BasicIntakeCommand;
-import frc.robot.commands.DefaultDriveCommand;
-import frc.robot.commands.ShootToAmpCommand;
-import frc.robot.commands.ShootToSpeakerCommand;
-import frc.robot.commands.SmartIntakeCommand;
+import frc.robot.commands.arm_commands.ArmIdleCommand;
+import frc.robot.commands.drive_commands.DefaultDriveCommand;
+import frc.robot.commands.intake_and_shooter.BasicIntakeCommand;
+import frc.robot.commands.intake_and_shooter.ShootToAmpCommand;
+import frc.robot.commands.intake_and_shooter.ShootToSpeakerCommand;
+import frc.robot.commands.intake_and_shooter.SmartIntakeCommand;
 import frc.robot.commands.led_commands.LEDIdleCommand;
 import frc.robot.simulationSystems.PhotonSim;
 import frc.robot.subsystems.ArmSubsystem;
