+++ src/main/java/frc/robot/commands/arm_commands/MoveArmToPositionCommand.java
@@ -1,4 +1,4 @@
-package frc.robot.commands;
+package frc.robot.commands.arm_commands;
 
 import edu.wpi.first.wpilibj2.command.ParallelRaceGroup;
 import frc.robot.subsystems.ArmSubsystem;
