+++ src/main/java/frc/robot/commands/arm_commands/MoveArmToShooterCommand.java
@@ -1,4 +1,4 @@
-package frc.robot.commands;
+package frc.robot.commands.arm_commands;
 
 import edu.wpi.first.wpilibj2.command.RunCommand;
 import frc.robot.subsystems.ArmSubsystem;
