+++ src/main/java/frc/robot/commands/drive_commands/DriveFacingShooter.java
@@ -1,4 +1,4 @@
-package frc.robot.commands;
+package frc.robot.commands.drive_commands;
 
 import edu.wpi.first.math.controller.PIDController;
 import edu.wpi.first.wpilibj2.command.Command;
