+++ src/main/java/frc/robot/commands/intake_and_shooter/BasicRunShooterCommand.java
@@ -1,4 +1,4 @@
-package frc.robot.commands;
+package frc.robot.commands.intake_and_shooter;
 
 import edu.wpi.first.wpilibj2.command.ParallelRaceGroup;
 import frc.robot.Constants.ShooterConstants;
