+++ src/main/java/frc/robot/subsystems/ArmSubsystem.java
@@ -14,7 +14,7 @@ import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
 import frc.robot.Constants.IntakeConstants;
 import frc.robot.Constants.IntakeConstants.ArmPIDConstants;
-import frc.robot.commands.MoveArmToPositionCommand;
+import frc.robot.commands.arm_commands.MoveArmToPositionCommand;
 import frc.utils.ExtraFunctions;
 import frc.utils.sim_utils.CANSparkMAXWrapped;
