+++ src/main/java/frc/robot/subsystems/IntakeSubsystem.java
@@ -11,8 +11,8 @@ import edu.wpi.first.wpilibj2.command.SubsystemBase;
 import edu.wpi.first.wpilibj2.command.WaitUntilCommand;
 import frc.robot.Constants.IntakeConstants;
 import frc.robot.Constants.IntakeConstants.ColorSensorConstants;
-import frc.robot.commands.LoadToShooterCommand;
-import frc.robot.commands.VibrateControllerCommand;
+import frc.robot.commands.intake_and_shooter.LoadToShooterCommand;
+import frc.robot.commands.util_commands.VibrateControllerCommand;
 import frc.utils.sim_utils.CANSparkMAXWrapped;
 import frc.utils.sim_utils.ColorSensorV3Wrapped;
