+++ src/test/java/command_tests/arm_tests/BasicIntakeTest.java
@@ -7,7 +7,7 @@ import static subsystem_tests.led_tests.utils.LEDTestUtils.checkForColorInAll;
 import command_tests.utils.CommandTestBase;
 import edu.wpi.first.wpilibj.Timer;
 import edu.wpi.first.wpilibj.util.Color;
-import frc.robot.commands.BasicIntakeCommand;
+import frc.robot.commands.intake_and_shooter.BasicIntakeCommand;
 import frc.robot.subsystems.IntakeSubsystem;
 import frc.utils.sim_utils.ColorSensorV3Wrapped;
 import org.junit.jupiter.api.AfterEach;
