+++ src/test/java/command_tests/drive_tests/DefaultDriveCommandTest.java
@@ -12,7 +12,7 @@ import edu.wpi.first.wpilibj.simulation.SimHooks;
 import edu.wpi.first.wpilibj.simulation.XboxControllerSim;
 import frc.robot.Constants.DriveConstants;
 import frc.robot.Constants.OIConstants;
-import frc.robot.commands.DefaultDriveCommand;
+import frc.robot.commands.drive_commands.DefaultDriveCommand;
 import frc.robot.subsystems.DriveSubsystem;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
